% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curry.r
\name{curry}
\alias{curry}
\title{Curry a function.}
\usage{
curry(func)
}
\arguments{
\item{func}{The function to curry.}
}
\value{
A curried version of \code{func}.
}
\description{
Creates a curried version of \code{func} that executes upon a parameterless
call.
}
\examples{

# Curry a function with labeled parameters. Note that a parameterless call
# must be made in order to execute the function.
sd_rm <- curry(sd)(na.rm = TRUE)
sd_rm(c(1, 2, 3, NA))()

# Curry a function with an ellipsis.
add_1_and_3_to <- curry(sum)(1, 3)
add_1_and_3_to(70)()

}
