% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_management.r
\name{device_management}
\alias{device_management}
\alias{cudaGetDeviceCount}
\alias{cudaGetDevice}
\alias{cudaSetDevice}
\alias{cudaDeviceReset}
\title{CUDA Device Management}
\usage{
cudaGetDeviceCount()

cudaGetDevice()

cudaSetDevice(device)

cudaDeviceReset()
}
\arguments{
\item{device}{A non-negative integer corresponding to the GPU you want to use.}
}
\description{
CUDA device management utilities.
}
\details{
\code{cudaGetDeviceCount()} returns the number of available GPU's.

\code{cudaGetDevice()} returns the device number (0, 1, ...) currently in use.

\code{cudaSetDevice()} sets the device to the supplied number (0, 1, ...).

\code{cudaDeviceReset()} Destroys all allocations and resets state on the
current device in the current process.
}
