% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_management.r
\name{device_management}
\alias{device_management}
\alias{cudaGetDeviceCount}
\alias{cudaGetDevice}
\alias{cudaSetDevice}
\alias{cudaDeviceReset}
\title{CUDA Device Management}
\usage{
cudaGetDeviceCount()

cudaGetDevice()

cudaSetDevice(device)

cudaDeviceReset()
}
\description{
CUDA device management utilities.
}
