% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_management.r
\name{memory_management}
\alias{memory_management}
\alias{cudaMalloc}
\alias{cudaFree}
\alias{cudaMemGetInfo}
\alias{cudaMemcpy}
\alias{cudaMemset}
\title{CUDA Memory Management}
\usage{
cudaMalloc(count, size)

cudaFree(dev_ptr)

cudaMemGetInfo()

cudaMemcpy(dst, src, count, size, kind)

cudaMemset(dev_ptr, value, count, size)
}
\description{
CUDA memory management utilities.
}
